requirements.txt
Django==2.1.12
djangorestframework==3.10.3
django-redis==4.10.0
psycopg2-binary==2.8.3
celery==4.3.0
uWSGI==2.0.18
requests==2.22.0
tenacity==5.1.1
PyJWT==1.7.1
channels==2.3.0
channels-redis==2.4.0
asgiref==3.2.2